requirements.txt
python-telegram-bot==22.0
imapclient==3.0.1
apscheduler==3.11.0
python-dotenv==1.1.0
beautifulsoup4==4.13.3